--- a/katsdptelstate/__init__.py
+++ b/katsdptelstate/__init__.py
@@ -16,9 +16,10 @@
 
 import warnings
 
-from .telescope_state import (TelescopeState, ConnectionError, InvalidKeyError,
-                              ImmutableKeyError, TimeoutError, CancelledError,
-                              DecodeError, EncodeError, RdbParseError,
+from .errors import (ConnectionError, InvalidKeyError,                  # noqa: F401
+                     ImmutableKeyError, TimeoutError, CancelledError,
+                     DecodeError, EncodeError, RdbParseError)
+from .telescope_state import (TelescopeState,                           # noqa: F401
                               PICKLE_PROTOCOL, encode_value, decode_value, set_allow_pickle,
                               ALLOWED_ENCODINGS, ENCODING_DEFAULT,
                               ENCODING_PICKLE, ENCODING_MSGPACK)