--- a/katsdptelstate/compat.py
+++ b/katsdptelstate/compat.py
@@ -14,7 +14,6 @@
 # limitations under the License.
 ################################################################################
 
-from __future__ import print_function, division, absolute_import
 
 import itertools