--- a/katsdptelstate/compat.py
+++ b/katsdptelstate/compat.py
@@ -14,7 +14,6 @@
 # limitations under the License.
 ################################################################################
 
-
 import itertools
 
 import redis