This is mostly in preparation to add asyncio support, but some of it is just nice to have.

- Python 2 support is dropped, Python 3.5 is required (I was going to go to 3.6, but we still have a lot of Ubuntu 16.04 machines around).
- Some of the more complex logic is moved into Lua scripts that run on the server. This was done partly because the same Lua script can then be used by the async version of the code, reducing duplication, but it's something I've been wanting to do for a while anyway because it cuts down the number of round trips and because a Lua script runs atomically, removing some race conditions. As a quick benchmark, adding 100 000 values to a sensor (with a local redis server) goes from about 12.5s -> 8s, but reading them all back (with a single `get_range`) increases from 0.17s to 0.23s. Improvements will probably be better with a remote redis server.
- As part of the Lua conversion, `get` (to fetch either an immutable or the latest value of a mutable) became a backend operation.
- The monolithic telescope_state.py is split up into separate modules. The goal is to get common code that can be reused for asyncio support out of the telescope_state.py module.
- flake8 cleanup (sorry @sratcliffe, your under-commenting had to go).

There's a lot here and a lot of it is mechanical stuff like moving code to other modules. I've tried to keep the individual commits meaningful so it may be easier to review a commit at a time.