@ajoubertza 2020-04-02
@bmerry Will you make a branch or tag with the last code to support Python 2 before merging this?  CAM will still need the Python 2 version for a while (currently installing from master).