@bmerry 2020-04-03
> @bmerry Will you make a branch or tag with the last code to support Python 2 before merging this? CAM will still need the Python 2 version for a while (currently installing from master).

Thanks for the reminder. I've just created a 'python2' branch of katsdptelstate from master that I expect will probably not get updates.