@ludwigschwardt 2020-04-03
The plan is also to make a Python 2 compatible release before merge that can be used for pinning if needed.