@bmerry 2020-04-21
Just noticed that the fixes I made didn't get pushed. @sratcliffe this should be ready for another review now.