@bmerry 2020-04-21
@ajoubertza before I merge, has CAM switched to using the python2 branch of katsdptelstate?