@ajoubertza 2020-04-21
> @ajoubertza before I merge, has CAM switched to using the python2 branch of katsdptelstate?

Yes, @bmerry.  Future deployments will use the new branch (and current deployments are unaffected).